requirements.txt
loguru>=0.7.3
fastapi[standard]>=0.115.12
uvicorn>=0.34.3
sqlalchemy>=2.0.41
sqlalchemy-utils>=0.41.2
nanoid
gpustat
PyJWT>=2.7.0
pytz>=2023.3
netspresso_trainer==1.4.0
boto3==1.38.35
requests_toolbelt
pymysql
aenum
celery
pika
pydantic_settings
better-exceptions==0.3.3
